+++ packages/components/src/components/select-v2/blocks/SelectEmpty.vue
@@ -6,6 +6,6 @@ const selectContext = injectSelectContext()
 
 <template>
   <div v-if="selectContext.isEmpty.value">
-    Empty
+    <slot />
   </div>
 </template>